Fixed two bugs
1. When storing the information of feedback terms in each documents,
the assignmentTermWeights method only create and set the values of a
FbTerm object once previously (just for the first doc it appears). It
is fixed by updating the values even when it exists in the
feedbackTermMap already.

2. Positions vectors for query terms can be null, so there is a
nullpointerexception when calculating its weight when it doesn't appear
in a feedback doc. Fixed by setting its weight to a smoothed
collection-based weight in the probQueryAtPos method. Normal feedback
terms can also have null positions vectors in some feedback docs, but
its weight in these docs will not be calculated, so it doesn't matter.